preparation/requirements.txt
django 
asgiref
sqlparse
djangorestframework